NBS: Remove need to manually manage chunkSource lifetime
For file-backed NBS, it's important that chunkSources get closed at
some point, to free up limited resources like file descriptors.
Previously, whenever a caller compacted or rebased a tableSet, the
caller had to manually close the chunkSources that were no longer
needed. This complicates logic around retrying in the event of an
optimistic lock failure.

This patch introduces a cache for open file descriptors, which
the fsTablePersister uses to manage the lifetime of files
automatically. The fdCache ref-counts open file descriptors,
but doesn't keep a hard cap on the number of open files. Once
the cache's target size is exceeded, opening a new file causes
the cache to look for fds with zero refs to close. It closes
as many of those as needed to get back under the target size.

Towards #3462