NBS: Tolerate other clients writing novel tables during compaction
Currently, we handle NBS compactions by having clients detect that one
is necessary, perform the compaction locally, and then attempt to
persist the work. The old strategy, removed in this patch, was to
abandon the work if any other client had changed anything upstream
since the compaction began. Since compaction can be slow, this
happened frequently. The new strategy is to tolerate new tables
appearing upstream as long as all the tables the client compacted
still exist. In that case, the client will now rebase against upstream
successfully, droping the tables that got compacted, and then attempt
to land the compaction once again.

Fixes #3462