NBS: Deal with concurrent file opens in fdCache
There was a race in the previous code. On a cache-miss, fdCache would
release its mutex and go os.Open() a file. Once it opened
successfully, fdCache would re-acquire the lock and update the
cache. The problem was that multiple concurrent cache-missers would
open the file and then all stomp on each other as they update the
cache one-by-one.

The problem here is that we don't want to serialize all file opens by
holding the lock while calling os.Open(). This version just checks the
cache again after opening the file and, if there's now a cache hit it
just drops the fd it just opened and refs the one in the cache.

Fixes #3473