NBS: Prevent concurrent loading of the same table index
By adding per-entry locking to indexCache, this patch allows both NBS
storage backends to ensure that only one goroutine is loading a given
table index at a time. Previoudly, if multiple readers tried to open
the same table at the same time, it was possible for both to get a
cache-miss and then read in the table index at the same time, which
wastes memory.  Now, one will win the race, load the index, and then
the other will see a cache hit.

Fixes #3581