csv-invert: Perf improvments!
Avoiding unneeded reallocation of a slice in sequence_chunker.go saves
about 15% of the runtime (a bit under a minute of the 6 minutes spent
inverting a list of about 1.5B entries). Tweaking the main loop that
processes list entries to avoid some map assignments, lookups, and
allocations saves another 15% or so, resulting in an overall savings
of about 2m on the 6m runtime of our test workload (as run on my
laptop).

Towards #3690