Enable walking refs in a Chunk without decoding to a Value
This patch introduces WalkRefs(c chunks.Chunk, cb RefCallback), which
allows the refs in a Chunk to be walked without first decoding the
Chunk into a Value. It's useful in e.g. datas.Pull(), which needs to
walk the Chunk graph, but doesn't actually care about the Values.

On my macbook, syncing e.g. a large dataset imported from a .csv file
is about 4x faster.