Add a multi-stage-build dockerfile
This is substantially based on https://hub.docker.com/r/scjalliance/noms/
but has been updated in two ways:

 1. It uses a multi-stage build to substantially reduce the size of the
    resultant image. The naive build image clocks in at ~1gb; the output
    image here is ~30mb.
 2. It takes advantage of the fact that the build context is already
    present, so we can speed up the build process by simply copying
    the context instead of cloning the data again.

This is intended to be compatible with #3790, and does not address
any of the issues mentioned there except simple creation of a dockerfile.
If the source repository for that PR were publicly available, this PR
would have targeted that branch instead of master.