Simplify json_import
Now that we have simple cli manipulation tools, we no longer need
the import tools to support committing on their own. You can just
import, get a hash, and use that to do what you want with.

Also updated json_import to use kingpin so the cli docs are
better.